+++ ui/zenoedit/comctrl/gv/zdictpanel.cpp
@@ -1,11 +1,11 @@
 #include "zdictpanel.h"
 #include "zsocketlayout.h"
-#include "../../render/renderparam.h"
-#include "../../style/zenostyle.h"
-#include "zenosocketitem.h"
+#include <zenoui/render/renderparam.h>
+#include <zenoui/style/zenostyle.h>
+#include <zenoui/comctrl/gv/zenosocketitem.h>
 #include "variantptr.h"
 #include <zenomodel/include/modelrole.h>
-#include "zenoparamwidget.h"
+#include <zenoui/comctrl/gv/zenoparamwidget.h>
 #include <zenomodel/include/uihelper.h>
 #include "zitemfactory.h"
 #include "zenogvhelper.h"
