+++ ui/zenoedit/comctrl/gv/zdictpanel.h
@@ -1,8 +1,8 @@
 #ifndef __ZDICT_PANEL_H__
 #define __ZDICT_PANEL_H__
 
-#include "zlayoutbackground.h"
-#include "callbackdef.h"
+#include <zenoui/comctrl/gv/zlayoutbackground.h>
+#include <zenoui/comctrl/gv/callbackdef.h>
 
 class ZDictSocketLayout;
 class ZenoParamPushButton;
