+++ ui/zenoedit/comctrl/gv/zsocketlayout.h
@@ -1,9 +1,9 @@
 #ifndef __ZSOCKET_LAYOUT_H__
 #define __ZSOCKET_LAYOUT_H__
 
-#include "zgraphicslayout.h"
-#include "zgraphicstextitem.h"
-#include "callbackdef.h"
+#include <zenoui/comctrl/gv/zgraphicslayout.h>
+#include <zenoui/comctrl/gv/zgraphicstextitem.h>
+#include <zenoui/comctrl/gv/callbackdef.h>
 
 class ZenoSocketItem;
 class ZSimpleTextItem;
