+++ ui/zenoedit/comctrl/zveceditor.h
