+++ ui/zenoui/comctrl/gv/zenoparamwidget.h
@@ -8,7 +8,6 @@
 #include "zenosocketitem.h"
 #include "zgraphicstextitem.h"
 #include <zenoui/comctrl/zcombobox.h>
-#include <zenoui/comctrl/zveceditor.h>
 #include <zenoui/comctrl/zcheckboxbar.h>
 #include <zenoui/comctrl/zcheckbox.h>
 #include <zenoui/comctrl/zlineedit.h>
