+++ ui/zenoui/comctrl/gv/zgraphicsnumslideritem.cpp
@@ -2,7 +2,6 @@
 #include <zenoui/style/zenostyle.h>
 #include <zeno/utils/log.h>
 #include "zgraphicstextitem.h"
-#include <zenoedit/zenoapplication.h>
 
 
 ZGraphicsNumSliderItem::ZGraphicsNumSliderItem(const QVector<qreal>& steps, QGraphicsItem* parent)
@@ -10,7 +9,7 @@ ZGraphicsNumSliderItem::ZGraphicsNumSliderItem(const QVector<qreal>& steps, QGra
     , m_steps(steps)
 {
     qreal maxWidth = 0, maxHeight = 0;
-    QFont font = zenoApp->font();
+    QFont font = QApplication::font();
     int padding = ZenoStyle::dpiScaled(5);
 
     for (int i = 0; i < m_steps.length(); i++)
