+++ ui/zenoui/comctrl/gv/zlineedititem.cpp
@@ -1,5 +1,4 @@
 #include "zlineedititem.h"
-#include <zenoedit/zenoapplication.h>
 
 
 ZLineEditItem::ZLineEditItem(QGraphicsItem* parent)
@@ -12,7 +11,7 @@ ZLineEditItem::ZLineEditItem(QGraphicsItem* parent)
     clr.setAlphaF(0.4);
     setDefaultTextColor(clr);
 
-    QFont font = zenoApp->font();
+    QFont font = QApplication::font();
     font.setPointSize(10);
     setFont(font);
     setMargins(15, 8, 15, 8);
@@ -29,7 +28,7 @@ ZLineEditItem::ZLineEditItem(const QString& text, QGraphicsItem* parent)
     setDefaultTextColor(clr);
 
 
-    QFont font = zenoApp->font();
+    QFont font = QApplication::font();
     font.setPointSize(10);    setFont(font);
     setMargins(15, 8, 15, 8);
 }
