+++ ui/zenoui/comctrl/zlineedit.h
@@ -43,23 +43,4 @@ private:
     bool m_bIconHover;
 };
 
-struct CURVE_DATA;
-class ZFloatLineEdit : public ZLineEdit {
-    Q_OBJECT
-  public:
-    explicit ZFloatLineEdit(QWidget *parent = nullptr);
-    explicit ZFloatLineEdit(const QString &text, QWidget *parent = nullptr);
-    void updateCurveData();
-
-  protected:
-    bool event(QEvent *event) override;
-
-  private slots:
-    void updateBackgroundProp(int frame);
-    void onUpdate(bool gl, int frame);
-
-  private:
-    ZTimeline *getTimeline();
-};
-
 #endif
